refactor: Move utils, type, and summary functions
- Move functions to separate files for better organization
