+++ cmd/cg/main.go
@@ -8,7 +8,6 @@ import (
 	"io"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"regexp"
 	"slices"
@@ -21,95 +20,6 @@ import (
 
 	"golang.org/x/mod/semver"
 )
-
-const (
-	vulnsURL = "https://vuln.go.dev"
-)
-
-type Job struct {
-	Package string
-	Symbols []string
-	Dir     string
-	Files   []string
-}
-
-type AffectedImportsDetails struct {
-	Symbols      []string
-	Type         string
-	FixedVersion []string
-}
-
-type UsedImportsDetails struct {
-	Symbols        []string `json:"Symbols,omitempty"`
-	CurrentVersion string   `json:"CurrentVersion,omitempty"`
-	ReplaceVersion string   `json:"ReplaceVersion,omitempty"`
-	FixCommands    []string `json:"FixCommands,omitempty"`
-}
-
-type Result struct {
-	IsVulnerable    string
-	UsedImports     map[string]UsedImportsDetails
-	Files           map[string][][]string
-	AffectedImports map[string]AffectedImportsDetails
-	GoCVE           string
-	CVE             string
-	Repository      string
-	Branch          string
-	Directory       string
-	Errors          []string
-	mu              sync.Mutex
-	Summary         string
-}
-
-type VulnReport struct {
-	ID       string     `json:"id"`
-	Aliases  []string   `json:"aliases"`
-	Affected []Affected `json:"affected"`
-}
-
-type Affected struct {
-	Package           Package           `json:"package"`
-	Ranges            []Range           `json:"ranges"`
-	EcosystemSpecific EcosystemSpecific `json:"ecosystem_specific"`
-}
-
-type EcosystemSpecific struct {
-	Imports []Import `json:"imports"`
-}
-
-type Import struct {
-	Path    string   `json:"path"`
-	Symbols []string `json:"symbols"`
-}
-
-type Package struct {
-	Name      string `json:"name"`
-	Ecosystem string `json:"ecosystem"`
-}
-
-type Range struct {
-	Type   string  `json:"type"`
-	Events []Event `json:"events"`
-}
-
-type Event struct {
-	Introduced string `json:"introduced,omitempty"`
-	Fixed      string `json:"fixed,omitempty"`
-}
-
-type PathVersion struct {
-	Path    string `json:"path"`
-	Version string `json:"version,omitempty"`
-}
-
-type Replace struct {
-	Old PathVersion
-	New PathVersion
-}
-
-type GoModEdit struct {
-	Replace []Replace
-}
 
 func main() {
 	tools := []string{"go", "digraph", "callgraph", "git"}
@@ -671,284 +581,4 @@ func getModPath(pkg, dir string, result *Result) string {
 		return ""
 	}
 	return strings.TrimSpace(string(out))
-}
-
-func runCommand(dir string, command string, args ...string) ([]byte, error) {
-	cmd := exec.Command(command, args...)
-	cmd.Env = append(os.Environ(), "GOFLAGS=-mod=mod", "GOWORK=off")
-	cmd.Dir = dir
-	out, err := cmd.CombinedOutput()
-	return out, err
-}
-
-func validateTools(tools []string) bool {
-	allAvailable := true
-	for _, tool := range tools {
-		_, err := exec.LookPath(tool)
-		if err != nil {
-			allAvailable = false
-			fmt.Printf("Failed finding %s package: %s", tool, err)
-		}
-	}
-	return allAvailable
-}
-
-func getGitBranch(result *Result) {
-	cmd := "git"
-	args := []string{"rev-parse", "--abbrev-ref", "HEAD"}
-	out, err := runCommand(result.Directory, cmd, args...)
-	if err != nil {
-		errMsg := fmt.Sprintf("Failed to run %s %s in %s: %s", cmd, strings.Join(args, " "), result.Directory, strings.TrimSpace(string(out)))
-		result.Errors = append(result.Errors, errMsg)
-	}
-	result.Branch = strings.TrimSpace(string(out))
-}
-
-func getGitURL(result *Result) {
-	cmd := "git"
-	args := []string{"remote", "get-url", "origin"}
-	out, err := runCommand(result.Directory, cmd, args...)
-	if err != nil {
-		errMsg := fmt.Sprintf("Failed to run %s %s in %s: %s", cmd, strings.Join(args, " "), result.Directory, strings.TrimSpace(string(out)))
-		result.Errors = append(result.Errors, errMsg)
-	}
-	result.Repository = strings.TrimSpace(string(out))
-}
-
-func formatIntroducedFixed(events []Event) []string {
-	var result []string
-	var introduced string
-
-	for _, e := range events {
-		if e.Introduced != "" {
-			introduced = e.Introduced
-		}
-		if e.Fixed != "" && introduced != "" {
-			pair := fmt.Sprintf("Introduced in %s and fixed in %s", introduced, e.Fixed)
-			result = append(result, pair)
-			introduced = ""
-		}
-	}
-
-	if introduced != "" {
-		result = append(result, fmt.Sprintf("Introdued in %s - ", introduced))
-	}
-
-	return result
-}
-
-func extractFormattedFixedVersions(inputs []string) []string {
-	re := regexp.MustCompile(`fixed in ([0-9a-zA-Z.\-]+)`)
-
-	var fixedVersions []string
-
-	for _, input := range inputs {
-		matches := re.FindAllStringSubmatch(input, -1)
-		for _, match := range matches {
-			if len(match) > 1 {
-				fixedVersions = append(fixedVersions, match[1])
-			}
-		}
-	}
-
-	return fixedVersions
-}
-
-func semVersion(v string) string {
-	if !strings.HasPrefix(v, "v") {
-		return "v" + v
-	}
-	return v
-}
-
-func generateSummaryWithGemini(result *Result) {
-	apiURL, apiKey, err := loadGeminiConfig()
-	if err != nil {
-		result.Summary = fmt.Sprintf("Failed to load Gemini config: %v", err)
-		return
-	}
-
-	prompt, err := BuildPrompt(result)
-	if err != nil {
-		result.Summary = fmt.Sprintf("Failed to build prompt: %v", err)
-		return
-	}
-
-	body := map[string]any{
-		"contents": []map[string]any{
-			{
-				"parts": []map[string]string{
-					{"text": "You are a senior software engineer specializing in Go security tools.\n" +
-						"You will be given a JSON output from a Go-based vulnerability scanner. Analyze and summarize the key findings clearly and concisely for inclusion in a security report.\n\n" + prompt},
-				},
-			},
-		},
-	}
-
-	jsonBody, err := json.Marshal(body)
-	if err != nil {
-		result.Summary = fmt.Sprintf("Failed to marshal Gemini request: %v", err)
-		return
-	}
-
-	fullURL := fmt.Sprintf("%s?key=%s", apiURL, apiKey)
-	req, err := http.NewRequest("POST", fullURL, bytes.NewReader(jsonBody))
-	if err != nil {
-		result.Summary = fmt.Sprintf("Failed to create Gemini request: %v", err)
-		return
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		result.Summary = fmt.Sprintf("Failed to connect to Gemini API: %v", err)
-		return
-	}
-	defer resp.Body.Close()
-
-	var response struct {
-		Candidates []struct {
-			Content struct {
-				Parts []struct {
-					Text string `json:"text"`
-				} `json:"parts"`
-			} `json:"content"`
-		} `json:"candidates"`
-	}
-
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		result.Summary = fmt.Sprintf("Failed to decode Gemini response: %v", err)
-		return
-	}
-
-	if len(response.Candidates) > 0 && len(response.Candidates[0].Content.Parts) > 0 {
-		result.Summary = response.Candidates[0].Content.Parts[0].Text
-	} else {
-		result.Summary = "No summary generated by Gemini."
-	}
-}
-
-func generateSummaryWithOllama(input string, result *Result) {
-	prompt, err := BuildPrompt(result)
-	body := map[string]any{
-		"model":       "llama3.2",
-		"prompt":      prompt,
-		"role":        "You are a senior software engineer specializing in Go security tools.",
-		"system":      "You will be given a JSON output from a Go-based vulnerability scanner. Analyze and summarize the key findings clearly and concisely for inclusion in a security report.",
-		"temperature": 0.9,
-		"max_tokens":  900,
-		"stream":      false,
-	}
-
-	jsonBody, err := json.Marshal(body)
-	if err != nil {
-		errMsg := fmt.Sprintf("Failed to marshal ollama request: %v\n", err)
-		result.Errors = append(result.Errors, errMsg)
-
-	}
-
-	resp, err := http.Post("http://localhost:11434/api/generate", "application/json", bytes.NewReader(jsonBody))
-	if err != nil {
-		errMsg := fmt.Sprintf("Failed to connect ollama API: %v\n", err)
-		result.Errors = append(result.Errors, errMsg)
-
-	}
-	defer resp.Body.Close()
-
-	var response struct {
-		Response string `json:"response"`
-	}
-
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		errMsg := fmt.Sprintf("Failed to decode ollama result: %v\n", err)
-		result.Errors = append(result.Errors, errMsg)
-	}
-
-	result.Summary = response.Response
-}
-
-func BuildPrompt(result *Result) (string, error) {
-
-	promptResult := &Result{
-		IsVulnerable: result.IsVulnerable,
-		UsedImports:  result.UsedImports,
-		GoCVE:        result.GoCVE,
-		CVE:          result.CVE,
-		Repository:   result.Repository,
-		Branch:       result.Branch,
-		Errors:       result.Errors,
-	}
-	resultJson, err := json.MarshalIndent(promptResult, "", "  ")
-	if err != nil {
-		return "", err
-	}
-
-	var sb strings.Builder
-
-	sb.WriteString(`You are a senior software engineer specializing in Go security tools.
-
-I have a JSON that represents the output of a Go-based vulnerability scanner. I want you to summarize its content for report. Reply only with the markdown without triple backticks. Be straight to the point in an elaborative manner.
-
-This JSON contains:
-- IsVulnerable: whether the project is affected.
-- UsedImports: import packages and how they are used in codebase.
-- Files: where symbols occur.
-- AffectedImports: vulnerable symbols and fixed versions of the CVE ID.
-- GoCVE: Go vulnerability ID.
-- CVE: General vulnerability ID.
-- Repository/Branch/Directory: context of the scanned code.
-- Errors: scanning issues if any.
-
-Here is JSON result after the scanning:
-
-`)
-	sb.WriteString("```json\n")
-	sb.Write(resultJson)
-	sb.WriteString("\n```\n\n")
-
-	sb.WriteString(`Please provide a concise and clear summary for technical and security teams including:
-- Whether the project is vulnerable and why.
-  - If yes
-    - If any field is null, ignore it.
-    - Which symbols and imports are involved only if the code is vulnerable.
-    - Which version should be used to fix it and how.
-  - If no,
-    - Simply say 'No vulnerability'.
-- Any errors or issues to be addressed.`)
-
-	return sb.String(), nil
-}
-
-func loadGeminiConfig() (url string, key string, err error) {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		return "", "", fmt.Errorf("could not find user home directory: %w", err)
-	}
-
-	confPath := filepath.Join(homeDir, ".gemini.conf")
-	file, err := os.Open(confPath)
-	if err != nil {
-		return "", "", fmt.Errorf("could not open config file: %w", err)
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.HasPrefix(line, "API_URL=") {
-			url = strings.TrimPrefix(line, "API_URL=")
-		} else if strings.HasPrefix(line, "API_KEY=") {
-			key = strings.TrimPrefix(line, "API_KEY=")
-		}
-	}
-
-	if err := scanner.Err(); err != nil {
-		return "", "", fmt.Errorf("error reading config file: %w", err)
-	}
-
-	if url == "" || key == "" {
-		return "", "", fmt.Errorf("API_URL or API_KEY not found in config")
-	}
-
-	return url, key, nil
 }
